+++ CHANGELOG.md
@@ -8,6 +8,8 @@ and this project adheres to
 
 ## [Unreleased]
 
+## [2.6.0] - 2026-09-23
+
 ### Added
 
 - `mpvd push -p`/`--playlist` flag to push tracks to an XSPF playlist file
@@ -158,7 +160,8 @@ and this project adheres to
 - `remove`/`rm` subcommand
 - `send` subcommand
 
-[unreleased]: https://github.com/rasmusmerzin/mpvd/compare/v2.5.0...main
+[unreleased]: https://github.com/rasmusmerzin/mpvd/compare/v2.6.0...main
+[2.6.0]: https://github.com/rasmusmerzin/mpvd/compare/v2.5.0...v2.6.0
 [2.5.0]: https://github.com/rasmusmerzin/mpvd/compare/v2.4.0...v2.5.0
 [2.4.0]: https://github.com/rasmusmerzin/mpvd/compare/v2.3.0...v2.4.0
 [2.3.0]: https://github.com/rasmusmerzin/mpvd/compare/v2.2.2...v2.3.0